Разбил main и исправил ошибки

+++ src/main/java/org/Dubrovsky/Aston/Homework/third/AdapterAndStrategy/APIs/StripeApi.java
@@ -1,8 +1,10 @@
 package org.Dubrovsky.Aston.Homework.third.AdapterAndStrategy.APIs;
 
+import java.math.BigDecimal;
+
 public class StripeApi {
 
-    public void makePayment(double amount) {
+    public void makePayment(BigDecimal amount) {
         System.out.println("Stripe: списано " + amount + "$");
     }
 }
