+++ src/main/java/org/Dubrovsky/Aston/Homework/third/AdapterAndStrategy/APIs/TBankApi.java
@@ -1,8 +1,10 @@
 package org.Dubrovsky.Aston.Homework.third.AdapterAndStrategy.APIs;
 
+import java.math.BigDecimal;
+
 public class TBankApi {
 
-    public void createTransaction(long amountInCents) {
+    public void createTransaction(BigDecimal amountInCents) {
 
         System.out.println("Т-Банк: операция на " + amountInCents + " копеек");
     }
