+++ src/main/java/org/Dubrovsky/Aston/Homework/third/AdapterAndStrategy/Adapters/PaypalAdapter.java
@@ -3,6 +3,8 @@ package org.Dubrovsky.Aston.Homework.third.AdapterAndStrategy.Adapters;
 import org.Dubrovsky.Aston.Homework.third.AdapterAndStrategy.APIs.PaypalApi;
 import org.Dubrovsky.Aston.Homework.third.AdapterAndStrategy.PaymentStrategy;
 
+import java.math.BigDecimal;
+
 public class PaypalAdapter implements PaymentStrategy {
 
     private final PaypalApi paypalApi;
@@ -12,7 +14,7 @@ public class PaypalAdapter implements PaymentStrategy {
     }
 
     @Override
-    public void pay(double amount) {
+    public void pay(BigDecimal amount) {
         paypalApi.sendMoney("USD", amount);
     }
 }
